+++ src/features/race/removeRepentanceDoor.ts
@@ -41,20 +41,27 @@ function checkRemoveRepentanceDoor() {
   }
 }
 
-// For seeded races, we remove all Repentance doors to mitigate seed divergence
-// (the exception is for races to Mother, since in that case, the Repentance path is forced)
-// (another exception is for races with a custom goal, since that goal could be in the Repentance
-// floors)
+/**
+ * For seeded races, we remove all Repentance doors to mitigate seed divergence. However, there are
+ * some exceptions:
+ *
+ * - Races to Mother, since in that case, the Repentance path is forced.
+ * - Races with a custom goal, since that goal could be in the Repentance floors.
+ * - Solo races, since we don't care about seed divergence in this case.
+ */
 function shouldRemoveRepentanceDoorOnSeededRace() {
   return (
     g.race.format === RaceFormat.SEEDED &&
     g.race.goal !== RaceGoal.MOTHER &&
-    g.race.goal !== RaceGoal.CUSTOM
+    g.race.goal !== RaceGoal.CUSTOM &&
+    !g.race.solo
   );
 }
 
-// For races to The Beast, the player must go to Depths 2
-// Thus, we must prevent them from going to the Mausoleum floors by deleting the doors
+/**
+ * For races to The Beast, the player must go to Depths 2. Thus, we must prevent them from going to
+ * the Mausoleum floors by deleting the doors.
+ */
 function shouldRemoveRepentanceDoorOnBeastRace() {
   return g.race.goal === RaceGoal.THE_BEAST && getEffectiveStage() === 5;
 }